requirements.txt
kivy
python-bidi
arabic_reshaper